floatfield

+++ runproject/run/models.py
@@ -5,5 +5,5 @@ from django.db import models
 class Run(models.Model):
     name = models.CharField(max_length=100)
     dur = models.IntegerField(default = None)
-    distance = models.IntegerField(default = None)
+    distance = models.FloatField(default = None)
     category = models.CharField(max_length=100, default = None)
